refactor: create class player
